+++ components/frontend/src/source/SourceEntity.test.js
@@ -3,7 +3,7 @@ import { Table } from 'semantic-ui-react';
 import { mount } from 'enzyme';
 import { SourceEntity } from './SourceEntity';
 
-describe('<Source />', () => {
+describe('<SourceEntity />', () => {
     it('render', () => {
         const wrapper = mount(<Table><Table.Body><SourceEntity entity_attributes={["attr"]} entity={{key:"1"}} /></Table.Body></Table>);
         expect(wrapper.find("TableRow").hasClass("status_unknown"))