+++ components/frontend/src/source/source_entity_status.js
@@ -1,0 +1,7 @@
+export const source_entity_status_name = {
+    unconfirmed: "Unconfirmed",
+    confirmed: "Confirmed",
+    fixed: "Fixed",
+    false_positive: "False positive",
+    wont_fix: "Won't fix"
+}